✨ running all limit steps
