+++ config/selectionCfg_ETau_UL18.cfg
@@ -243,8 +243,8 @@ DNNoutSM_kl_1_jetdown11    = 20,0,1
 
 # Dummy binning needed to have make the binning for different categories work properly.
 # These binnings will actually be overwritten by:  category:variable = 0.0, ....
-HH_mass                 = 245.0, 255.0, 265.0, 275.0, 285.0, 315.0, 325.0, 375.0, 425.0, 475.0, 525.0, 575.0, 625.0, 675.0, 725.0, 775.0, 825.0, 875.0, 925.0, 1075.0, 1325.0, 1675.0, 1825.0, 2175.0, 2725.0, 3275.0
-HHKin_mass              = 245.0, 255.0, 265.0, 275.0, 285.0, 315.0, 325.0, 375.0, 425.0, 475.0, 525.0, 575.0, 625.0, 675.0, 725.0, 775.0, 825.0, 875.0, 925.0, 1075.0, 1325.0, 1675.0, 1825.0, 2175.0, 2725.0, 3275.0
+HH_mass = 245.0, 255.0, 265.0, 275.0, 285.0, 315.0, 325.0, 375.0, 425.0, 475.0, 525.0, 575.0, 625.0, 675.0, 725.0, 775.0, 825.0, 875.0, 925.0, 1075.0, 1325.0, 1675.0, 1825.0, 2175.0, 2725.0, 3275.0
+HHKin_mass = 245.0, 255.0, 265.0, 275.0, 285.0, 315.0, 325.0, 375.0, 425.0, 475.0, 525.0, 575.0, 625.0, 675.0, 725.0, 775.0, 825.0, 875.0, 925.0, 1075.0, 1325.0, 1675.0, 1825.0, 2175.0, 2725.0, 3275.0
 DNNoutSM_kl_1              = 0.0, 0.5, 1.0
 DNNoutSM_kl_1_tauup_DM0    = 0.0, 0.5, 1.0
 DNNoutSM_kl_1_taudown_DM0  = 0.0, 0.5, 1.0
